[gofonoext] Return NULL from set_mms_imsi() if mm is invalid
The caller should wait until OfonoExtModemManager becomes valid and
only then make the calls. It's not very convenient and we could queue
these early requests and submit them later but at this point it doesn't
look like it's worth the trouble. If necessary, queuing can be implemented
later without breaking the API.